Upgrade dependencies

+++ robot-system-lib-parent/robot-system-lib-core/src/main/java/org/arig/robot/filters/pid/AbstractPidFilter.java
@@ -96,7 +96,7 @@ public abstract class AbstractPidFilter implements IPidFilter {
     }
 
     public void setSampleTime(double value, TimeUnit unit) {
-        setSampleTimeMs((double) unit.toMillis((long) value));
+        setSampleTimeMs(unit.toMillis((long) value));
     }
 
     @Override
